ami render working
