+++ src/app/layout.tsx
@@ -1,24 +1,13 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import dynamic from 'next/dynamic'
 
 const inter = Inter({ subsets: ['latin'] })
-
-export const ami = dynamic(
-  () => {
-    return import("./services/visualization");
-  },
-  { ssr: false }
-)
 
 export const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
 }
-
-// https://unpkg.com/ami.js@next/build/ami.min.js
-//<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/ami.js/0.0.13/ami.min.js"></script> 
 
 export default function RootLayout({
   children,
@@ -31,10 +20,11 @@ export default function RootLayout({
         {/* Load THREEJS before AMI */}
         <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.7.3/dat.gui.min.js"></script>
         <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r119/three.js"></script>
-        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/ami.js/0.0.13/ami.min.js"></script>
-        <script src='app.js'></script>
+        <script type="text/javascript" src="https://unpkg.com/ami.js@next/build/ami.min.js"></script>
       </head>
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        {children}
+      </body>
     </html>
   )
 }
