+++ src/app/physician-view/page.tsx
@@ -2,18 +2,10 @@ import React from 'react';
 import dynamic from 'next/dynamic';
 import NavBar from '../navbar/navbar';
 import InfoPanel from '../info-panel/info-panel';
-//import Visualization from '../services/visualization';
-import Visualization from '../services/visualization_copy';
 import './physician-view.css';
 
-var img = require('../../assets/tumor.jpg');
-// var script = require('./testing.js');
-
-// https://sentry.io/answers/window-is-not-defined/
 const VisualizationComponent = dynamic(
-  //() => import('../services/visualization'),
-  () => import('../services/visualization_copy'),
-  //() => import('../services/visualization_c'),
+  () => import('../services/visualization'),
   {
     ssr: false,
   }
@@ -43,10 +35,3 @@ export default function Page() {
   )
 }
 
-/*
-<div>
-          <div id='my-gui-container'></div>
-          <div id='container'></div>
-          <VisualizationComponent></VisualizationComponent>
-        </div> 
-*/
