--- src/app/services/visualization_c.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-var script = require('./testing.js');
-
-const Visualization = () => {
-  return (
-    <script src={script}></script>
-  );
-};
-  
-export default Visualization;
-
